+++ public/index.php
@@ -413,6 +413,9 @@ if (\MuseDockPanel\Controllers\SetupController::needsSetup()) {
 \MuseDockPanel\Router::post('/mail/relay/domains/{id}/delete', 'MailController@relayDomainDelete');
 \MuseDockPanel\Router::post('/mail/relay/users/store', 'MailController@relayUserStore');
 \MuseDockPanel\Router::post('/mail/relay/users/{id}/delete', 'MailController@relayUserDelete');
+\MuseDockPanel\Router::post('/mail/relay/queue/flush', 'MailController@relayQueueFlush');
+\MuseDockPanel\Router::post('/mail/relay/queue/delete', 'MailController@relayQueueDelete');
+\MuseDockPanel\Router::post('/mail/relay/queue/delete-message', 'MailController@relayQueueDeleteMessage');
 \MuseDockPanel\Router::post('/mail/migrations/preflight', 'MailController@migrationPreflight');
 \MuseDockPanel\Router::post('/mail/migrations/relay/execute', 'MailController@migrationRelayExecute');
 \MuseDockPanel\Router::post('/mail/webmail/save', 'MailController@webmailSave');